your_project/requirements.txt
streamlit==1.32.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0 